frame-select/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.1.0
opencv-python>=4.8.1.78
numpy>=1.24.3
pydantic>=2.5.0
python-dotenv>=1.0.0
ultralytics>=8.0.0
scikit-learn>=1.3.0
imagehash>=4.3.1
torch>=2.0.0
torchvision>=0.15.0